services/billing-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
reportlab==4.0.7
prometheus-fastapi-instrumentator==6.1.0
